+++ mkdocs.yml
@@ -41,6 +41,7 @@ nav:
     - Overview: categorical/index.md
     - Numerical &times; Numerical Data: categorical/num_num.md
     - Categorical &times; Numerical Data: categorical/cat_num.md
+    - Aggregation: categorical/aggregation.md
   - Event Handling:
     - Overview: events/index.md
     - Canvas Events: events/canvas_events.md
